# perf(solc): parallelize dependency graph 