--- a/ethers-solc/Cargo.toml
+++ b/ethers-solc/Cargo.toml
@@ -40,6 +40,7 @@ rayon = "1.5.3"
 rand = { version = "0.8.5", optional = true }
 path-slash = "0.2.1"
 cfg-if = "1.0.0"
+parking_lot = "0.12.1"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
 home = "0.5.3"