--- a/ethers-solc/benches/resolve_graph.rs
+++ b/ethers-solc/benches/resolve_graph.rs
@@ -3,13 +3,10 @@
 extern crate criterion;
 
 use criterion::Criterion;
-use ethers_solc::Graph;
-use ethers_solc::project_util::TempProject;
+use ethers_solc::{project_util::TempProject, Graph, ProjectPathsConfig};
 
-fn resolve_graph_bench_benchmark(c: &mut Criterion) {
-    let tmp = TempProject::checkout("transmissions11/solmate").expect("failed to checkout project");
-    let paths = tmp.paths();
-    let mut group = c.benchmark_group("resolve graph");
+fn resolve(id: &str, paths: &ProjectPathsConfig, c: &mut Criterion) {
+    let mut group = c.benchmark_group(format!("resolve {}", id));
     group.bench_function("sequential", |b| {
         b.iter(|| {
             Graph::resolve_sync(paths).unwrap();
@@ -22,6 +19,20 @@ fn resolve_graph_bench_benchmark(c: &mut Criterion) {
     });
 }
 
+fn resolve_solmate_benchmark(c: &mut Criterion) {
+    let id = "transmissions11/solmate";
+    let tmp = TempProject::checkout(id).expect("failed to checkout project");
+    resolve(id, tmp.paths(), c)
+}
 
-criterion_group!(benches, resolve_graph_bench_benchmark);
+fn resolve_spells_benchmark(c: &mut Criterion) {
+    if let Ok(spells) = std::env::var("SPELLS_DIR") {
+        // cloning spells repo takes forever so this is only run when path to local dir is set
+        let id = "makerdao/spells-mainnet";
+        let paths = ProjectPathsConfig::dapptools(spells).unwrap();
+        resolve(id, &paths, c)
+    }
+}
+
+criterion_group!(benches, resolve_spells_benchmark, resolve_solmate_benchmark);
 criterion_main!(benches);