<!--
Thank you for your Pull Request. Please provide a description above and review
the requirements below.

Bug fixes and new features should include tests.

Contributors guide: https://github.com/gakonst/ethers-rs/blob/master/CONTRIBUTING.md

The contributors guide includes instructions for running rustfmt and building the
documentation.
-->

## Motivation
previously resolving the dependency graph was implemented sequentially

This resolves all dependencies in parallel
<!--
Explain the context and why you're making that change. What is the problem
you're trying to solve? In some cases there is not a problem and this can be
thought of as being the motivation for your change.
-->

## Solution
use rayon + locks to guard the relevant maps

for every new import, a new task is spawned in the threadpool's scope
once the scope finishes, all node data can be safely moved out of the locks.


obviously, this is more beneficial for larger projects, 
for solmate this is about ~10-15% (21ms -> 18ms) faster, for spells ~50% (43ms -> 22ms) but this is all in the ms ballpark. So actually not that big of a deal...

This still contains the old `_sync` variations of `Graph::resolve` which could be removed if we want to merge this.

<!--
Summarize the solution and provide any necessary context needed to understand
the code change.
-->

## PR Checklist

- [ ] Added Tests
- [ ] Added Documentation
- [ ] Updated the changelog